Improve rule matching in tests
There was a problem with the test for AssessRulesMP as the rules didn't
always return in the same order but the test was strict about the order.
Therefore I changed to test to check if the rules are the same but not
wory about the order.  To do this I improved the Sort method to sort by
lexical rule string order after everything else.  Along the way this
showed up a fault with LimitRuleAssessments, both in the accuracy of the
tests and in how many rules were being returned.  This is also now
corrected.